Allow to load screenshots in original size

+++ src/test/java/borg/edtrading/BodyScannerTest.java
@@ -1,7 +1,19 @@
 package borg.edtrading;
 
+import boofcv.core.image.ConvertImage;
+import boofcv.struct.image.GrayF32;
+import boofcv.struct.image.GrayU8;
+import boofcv.struct.image.ImageBase;
 import borg.edtrading.bodyscanner.BodyScanner;
 import borg.edtrading.bodyscanner.BodyScannerResult;
+import borg.edtrading.imagetransformation.Transformation;
+import borg.edtrading.imagetransformation.TransformationException;
+import borg.edtrading.imagetransformation.simple.RgbToGrayTransformation;
+import borg.edtrading.screenshots.Region;
+import borg.edtrading.screenshots.Screenshot;
+import borg.edtrading.templatematching.Match;
+import borg.edtrading.templatematching.Template;
+import borg.edtrading.templatematching.TemplateMatcher;
 import borg.edtrading.util.ImageUtil;
 import org.apache.commons.io.FileUtils;
 import org.apache.logging.log4j.LogManager;
@@ -17,6 +29,7 @@ import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 import java.util.Random;
+import java.util.stream.Collectors;
 
 import javax.imageio.ImageIO;
 
@@ -44,9 +57,9 @@ public class BodyScannerTest {
         //2016-09-29 08-16-28 Paul-Friedrichs Star
         //2016-10-03 08-37-57 Altair
         //2016-09-29 08-24-03 BD+63 1764
-        List<File> screenshotFiles = BodyScannerApp.selectSpecificScreenshot("2016-10-02 06-57-48 Sirius.png");
+        //List<File> screenshotFiles = BodyScannerApp.selectSpecificScreenshot("2016-10-02 06-57-48 Sirius.png");
         //List<File> screenshotFiles = BodyScannerApp.selectRandomScreenshot();
-        //List<File> screenshotFiles = BodyScannerApp.selectAllScreenshots();
+        List<File> screenshotFiles = BodyScannerApp.selectAllScreenshots();
         int n = 0;
         int total = screenshotFiles.size();
         int batchSize = 10;
@@ -65,12 +78,12 @@ public class BodyScannerTest {
 
             logger.trace("Testing " + screenshotFile.getName());
             BodyScanner scanner = new BodyScanner();
-            scanner.setDebugThresholdImage(true);
-            scanner.setDebugBlurredImage(true);
-            scanner.setDebugAlphanumTemplates(true);
-            scanner.setDebugAlphanumTextLines(true);
+            scanner.setDebugThresholdImage(false);
+            scanner.setDebugBlurredImage(false);
+            scanner.setDebugAlphanumTemplates(false);
+            scanner.setDebugAlphanumTextLines(false);
             scanner.setDebugAllTemplates(true);
-            scanner.setDebugAllTextLines(true);
+            scanner.setDebugAllTextLines(false);
             scanner.setDebugFinal(true);
             BodyScannerResult result = scanner.scanScreenshotFile(screenshotFile);
             if (result.getThresholdDebugImage() != null) {
@@ -102,6 +115,7 @@ public class BodyScannerTest {
     }
 
     static void copyLearnedChars(File learnedSetDir, File targetSetDir) throws IOException {
+        List<Template> verifiedTemplates = Template.fromFolder("BodyScanner");
         final Random random = new Random(System.currentTimeMillis());
         File[] templateTextDirs = learnedSetDir.listFiles(new FileFilter() {
             @Override
@@ -115,13 +129,31 @@ public class BodyScannerTest {
         });
         if (templateTextDirs != null) {
             for (File templateTextDir : templateTextDirs) {
+                List<Template> verifiedTemplatesForCurrentChar = verifiedTemplates.stream().filter(t -> t.getFile().getParentFile().getName().equals(templateTextDir.getName())).collect(Collectors.toList());
+
                 // Learn max 1 per type
                 File[] pngFiles = templateTextDir.listFiles(new FileFilter() {
                     @Override
                     public boolean accept(File file) {
-                        return file.getName().endsWith(".png");
+                        try {
+                            Region region = Screenshot.loadFromFile(file, -1, -1, null).getAsRegion();
+                            region.applyTransformation("GRAY", new RgbToGrayTransformation());
+                            region.applyTransformation("F32", new Transformation() {
+                                @Override
+                                public ImageBase<?> transform(ImageBase<?> inputImage) throws TransformationException {
+                                    return ConvertImage.convert((GrayU8) inputImage, (GrayF32) null);
+                                }
+                            });
+                            Match bestVerifiedMatch = new TemplateMatcher().bestMatchingTemplate(region, verifiedTemplatesForCurrentChar);
+                            return bestVerifiedMatch != null && bestVerifiedMatch.getErrorPerPixel() <= BodyScanner.ERROR_PER_PIXEL_GUESSED;
+                        } catch (IOException e) {
+                            return false;
+                        }
                     }
                 });
+                if (pngFiles != null && pngFiles.length == 0) {
+                    logger.warn("All learned chars differ too much from the currently known: " + templateTextDir);
+                }
                 if (pngFiles != null && pngFiles.length >= 1) {
                     File randomPngFile = pngFiles[random.nextInt(pngFiles.length)];
                     File targetDir = new File(targetSetDir, templateTextDir.getName());
@@ -142,7 +174,7 @@ public class BodyScannerTest {
                         }
                     });
                     for (int i = 0; i < learned.length; i++) {
-                        if (i >= 10) {
+                        if (i >= 10 && !learned[i].getName().startsWith("CRAP")) {
                             learned[i].delete();
                         }
                     }
